[WIP][RFC] tty: n_tty: introduce RING_BUFFER_SIZE macro
This macro helps avoid ring buffer size calculation mistakes or
unintentional arithmetic overflow.

TODO: Find a nice way to word this commit message. Really the
overflow/wrap-around is not "unintentional", the math works out and
developers are doing the arithmetic this way on purpose. I guess what is
meant by "unintentional" is that it upsets sanitizers "unintentionally".

Signed-off-by: Justin Stitt <justinstitt@google.com>
---
[WIP] because not sure if we need this macro. Even if we do, it should go
into a header file for broader use.