n_tty: use signed type to avoid overflow in chars_in_buffer
size_t is now defined as non-wrapping (should not overflow) in the
kernel. Certain sections of the tty layer rely on arithmetic overflow
using size_t, since Commit e30364c70895c1ba ("tty: n_tty: unify counts
to size_t").

There seems to not be a bug with this code as the math "works out" after
factoring in the wrap-around. This is no accident, developers have
intentionally opted into this behavior. However, the need for better
integer overflow sanitizer coverage means we should remove false
positives such as this one to achieve a better signal to noise ratio --
so we can catch _real_ bugs.

Signed-off-by: Justin Stitt <justinstitt@google.com>